+++ ui/src/components/Property/PropertyRead.vue
@@ -21,21 +21,19 @@
   import { useRoute } from 'vue-router';
   import { ElMessage } from 'element-plus';
   import { reactive, ref, watch, onMounted, getCurrentInstance } from 'vue';
-  import { RefreshRight, Search } from '@element-plus/icons-vue';
-  import { fetchProperties, deleteProperty } from '@/api/index';
+  import { RefreshRight, Search, TrendCharts } from '@element-plus/icons-vue';
+  import { fetchProperties } from '@/api/index';
   import { yamlToJson } from '@/utils/yaml';
   import CodeMirror from '@/components/CodeMirror/index.vue';
-  import PropertyEditor from './PropertyEditor.vue';
-  import PropertyValueReader from './PropertyValueReader.vue';
   import FormHeader from '../common/FormHeader.vue';
+  import TraceTree from '../TraceTree/TraceContent.vue';
+  import PropertyTable from '@/components/common/PropertyTable.vue';
 
   const { proxy } = getCurrentInstance();
   // Loading
   const route = useRoute();
   const $loadingCreate = getCurrentInstance().appContext.config.globalProperties.$loadingCreate;
   const $loadingClose = proxy.$loadingClose;
-  const propertyEditorRef = ref();
-  const propertyValueViewerRef = ref();
   const yamlRef = ref(null);
   const data = reactive({
     group: route.params.group,
@@ -44,6 +42,8 @@
   });
   const yamlCode = ref(`name: ${data.name}
 limit: 10`);
+  const showTracesDialog = ref(false);
+  const traceData = ref(null);
   const getProperties = async (params) => {
     $loadingCreate();
     const res = await fetchProperties({ groups: [data.group], name: data.name, limit: 10, ...params });
@@ -55,32 +55,12 @@ limit: 10`);
       });
       return;
     }
+    traceData.value = res.trace;
     data.tableData = (res.properties || []).map((item) => {
       item.tags.forEach((tag) => {
         tag.value = JSON.stringify(tag.value);
       });
       return item;
-    });
-  };
-  const openPropertyView = (data) => {
-    propertyValueViewerRef?.value.openDialog(data);
-  };
-
-  const ellipsizeValueData = (data) => {
-    return data.value.slice(0, 20) + '...';
-  };
-  const openEditField = (index) => {
-    const item = data.tableData[index];
-    const param = {
-      group: item.metadata.group,
-      name: item.metadata.name,
-      modRevision: item.metadata.modRevision,
-      createRevision: item.metadata.createRevision,
-      id: item.id,
-      tags: JSON.parse(JSON.stringify(item.tags)),
-    };
-    propertyEditorRef?.value.openDialog(true, param).then(() => {
-      getProperties();
     });
   };
 
@@ -101,24 +81,7 @@ limit: 10`);
         });
       });
   }
-  const deleteTableData = async (index) => {
-    const item = data.tableData[index];
-    $loadingCreate();
-    const response = await deleteProperty(item.metadata.group, item.metadata.name, item.id);
-    $loadingClose();
-    if (response.error) {
-      ElMessage({
-        message: `Failed to delete property: ${response.error.message}`,
-        type: 'error',
-      });
-      return;
-    }
-    ElMessage({
-      message: 'successed',
-      type: 'success',
-    });
-    getProperties();
-  };
+
   onMounted(() => {
     getProperties();
   });
@@ -141,58 +104,29 @@ limit: 10`;
         <FormHeader :fields="data" />
       </template>
       <div class="button-group-operator">
-        <div>
-          <el-button size="small" :icon="Search" @click="searchProperties" plain />
-          <el-button size="small" :icon="RefreshRight" @click="getProperties" plain />
-        </div>
+        <el-button size="small" :icon="Search" @click="searchProperties" plain />
+        <el-button size="small" :icon="RefreshRight" @click="getProperties" plain />
       </div>
       <CodeMirror ref="yamlRef" v-model="yamlCode" mode="yaml" style="height: 200px" :lint="true" />
-      <el-table :data="data.tableData" style="width: 100%; margin-top: 20px" border>
-        <el-table-column label="Group" prop="metadata.group" width="100"></el-table-column>
-        <el-table-column label="Name" prop="metadata.name" width="120"></el-table-column>
-        <el-table-column label="ModRevision" prop="metadata.modRevision" width="120"></el-table-column>
-        <el-table-column label="CreateRevision" prop="metadata.createRevision" width="140"></el-table-column>
-        <el-table-column label="ID" prop="id" width="150"></el-table-column>
-        <el-table-column label="Tags">
-          <template #default="scope">
-            <el-table :data="scope.row.tags">
-              <el-table-column label="Key" prop="key" width="150"></el-table-column>
-              <el-table-column label="Value" prop="value">
-                <template #default="scope">
-                  {{ ellipsizeValueData(scope.row) }}
-                  <el-button
-                    link
-                    type="primary"
-                    @click.prevent="openPropertyView(scope.row)"
-                    style="color: var(--color-main); font-weight: bold"
-                    >view</el-button
-                  >
-                </template>
-              </el-table-column>
-            </el-table>
-          </template>
-        </el-table-column>
-        <el-table-column label="Operator" width="150">
-          <template #default="scope">
-            <el-button
-              link
-              type="primary"
-              @click.prevent="openEditField(scope.$index)"
-              style="color: var(--color-main); font-weight: bold"
-              >Edit</el-button
-            >
-            <el-popconfirm @confirm="deleteTableData(scope.$index)" title="Are you sure to delete this?">
-              <template #reference>
-                <el-button link type="danger" style="color: red; font-weight: bold">Delete</el-button>
-              </template>
-            </el-popconfirm>
-          </template>
-        </el-table-column>
-      </el-table>
+      <div style="margin-top: 20px; margin-bottom: 10px; display: flex; justify-content: flex-end">
+        <el-button :icon="TrendCharts" @click="showTracesDialog = true" :disabled="!traceData" plain>
+          <span>Debug Trace</span>
+        </el-button>
+      </div>
+      <PropertyTable :data="data.tableData" :border="true" :show-operator="true" @refresh="getProperties" />
     </el-card>
-    <PropertyEditor ref="propertyEditorRef"></PropertyEditor>
-    <PropertyValueReader ref="propertyValueViewerRef"></PropertyValueReader>
   </div>
+  <el-dialog
+    v-model="showTracesDialog"
+    width="90%"
+    :destroy-on-close="true"
+    @closed="showTracesDialog = false"
+    class="trace-dialog"
+  >
+    <div style="max-height: 74vh; overflow-y: auto">
+      <TraceTree :trace="traceData" />
+    </div>
+  </el-dialog>
 </template>
 <style lang="scss" scoped>
   :deep(.el-card) {
@@ -202,7 +136,7 @@ limit: 10`;
   .button-group-operator {
     display: flex;
     flex-direction: row;
-    justify-content: space-between;
+    justify-content: end;
     margin-bottom: 10px;
   }
 </style>